🧹 Refactor large index.tsx in Hero directory by extracting sections into individual subcomponents.
