+++ src/components/free-for-charity-ffc-web-developer-training-guide-components/Hero/sections/Section6Wpmudev.tsx
@@ -1,0 +1,115 @@
+import React from 'react'
+
+const Section6Wpmudev = () => {
+  return (
+    <>
+      {/* 6. WPMUdev */}
+      <section id="section6" className="mb-6">
+        <ol className="list-decimal list-inside" start={6}>
+          <li className="text-[30px] md:text-[36px]  font-[700] text-[#005AA0] mt-16 mb-6 pb-3 border-b border-[#e1e5eb]">
+            WPMUdev (WordPress Plugins)
+          </li>
+        </ol>
+
+        <p className="text-[14px] font-[500] leading-[25px] text-[#333d47]">
+          WPMUdev extends your charity’s website functionality with a range of essential WordPress
+          plugins.
+        </p>
+
+        <h3 className="text-[24px] leading-[31px] font-[700] text-[#0066B8] mt-10 mb-4">
+          6.1. Setup Process
+        </h3>
+
+        <ul className="pl-[1.5rem] pb-[23px] list-disc space-y-[0.75rem]">
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <span className="font-[600] text-[#1c2a38]">Identify Useful Plugins: </span>
+            Install plugins from WPMUdev that boost site performance, security, SEO, and more.
+          </li>
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <span className="font-[600] text-[#1c2a38]">Configuration: </span>
+            Follow the installation and configuration instructions provided in the WPMUdev
+            dashboard.
+          </li>
+        </ul>
+
+        <h3 className="text-[24px] leading-[31px] font-[700] text-[#0066B8] mt-[3rem] mb-6">
+          6.2. Learning Resources
+        </h3>
+
+        <p className="leading-[26px] mb-[0.75rem] text-[14px] font-[500] text-[#333d47]">
+          WPMUdev provides extensive documentation for each of its pro plugins. Here are the direct
+          links to the documentation for their most popular tools:
+        </p>
+
+        <ul className="pl-[1.5rem] pb-[23px] list-disc space-y-[0.75rem]">
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <a
+              href="https://wpmudev.com/docs/wpmu-dev-plugins/smush/"
+              className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+            >
+              Smush Pro (Image Optimization)
+            </a>
+          </li>
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <a
+              href="https://wpmudev.com/docs/wpmu-dev-plugins/hummingbird/"
+              className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+            >
+              Hummingbird Pro (Performance & Caching)
+            </a>
+          </li>
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <a
+              href="https://wpmudev.com/docs/wpmu-dev-plugins/defender/"
+              className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+            >
+              Defender Pro (Security)
+            </a>
+          </li>
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <a
+              href="https://wpmudev.com/docs/wpmu-dev-plugins/forminator/"
+              className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+            >
+              Forminator Pro (Forms & Surveys)
+            </a>
+          </li>
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <a
+              href="https://wpmudev.com/docs/wpmu-dev-plugins/snapshot/"
+              className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+            >
+              Snapshot Pro (Backups)
+            </a>
+          </li>
+          <li className="leading-[26px] pl-[0.5rem] text-[14px] font-[500] text-[#333d47]">
+            <a
+              href="https://wpmudev.com/docs/wpmu-dev-plugins/branda/"
+              className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+            >
+              Branda Pro (White Labeling)
+            </a>
+          </li>
+        </ul>
+        <p className="leading-[26px] mb-[0.75rem] text-[14px] font-[500] text-[#333d47]">
+          For general tutorials and community support, you can also visit the{' '}
+          <a
+            href="https://wpmudev.com/blog/"
+            className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+          >
+            WPMUdev Blog
+          </a>
+          and their
+          <a
+            href="https://wpmudev.com/forums/forum/support/"
+            className="text-[#0066B8] underline hover:text-[#005AA0] transition-colors duration-200"
+          >
+            Community Forums.
+          </a>
+        </p>
+      </section>
+    </>
+  )
+}
+
+export default Section6Wpmudev
