+++ src/components/free-for-charity-ffc-web-developer-training-guide-components/Hero/sections/Section7MicrosoftClarity.tsx
@@ -1,0 +1,110 @@
+import React from 'react'
+
+const Section7MicrosoftClarity = () => {
+  return (
+    <>
+      {/* 7. Microsoft Clarity */}
+      <section id="section7" className="mb-6">
+        <ol className="list-decimal list-inside" start={7}>
+          <li className="text-[30px] md:text-[36px]  font-[700] text-[#005AA0] mt-16 mb-6 pb-3 border-b border-[#e1e5eb]">
+            Microsoft Clarity (Analytics)
+          </li>
+        </ol>
+
+        <p className="text-[14px] font-[500] leading-[25px] text-[#333d47]">
+          Microsoft Clarity provides insights into how users interact with the charity website so
+          you can optimize its performance.
+        </p>
+
+        <h3 className="text-[24px] leading-[31px] font-[700] text-[#0066B8] mt-10 mb-4">
+          7.1. Setup Process
+        </h3>
+
+        <ul className="pl-[1.5rem] p-[0_0_23px_1em] list-disc">
+          <li className="pl-[0.5rem] mb-[0.75rem] text-[14px] font-[500] text-[#333d47]">
+            <span className="font-[700]">Install Plugin or Embed Script: </span>
+            Add the Microsoft Clarity plugin via WordPress or embed the tracking script in the
+            website’s header.
+          </li>
+          <li className="pl-[0.5rem] mb-[0.75rem] text-[14px] font-[500] text-[#333d47]">
+            <span className="font-[700]">Monitor and Analyze User Behavior: </span>
+            Use the dashboard to view heatmaps, session recordings, and other insights.
+          </li>
+        </ul>
+
+        <h3 className="text-[24px] leading-[31px] font-[700] text-[#0066B8] mt-[3rem] mb-6">
+          7.2. Learning Resources
+        </h3>
+
+        <h4 className="text-[#1c2a38] text-[18px] pb-[10px] font-[700] leading-[23px] mt-[1rem]">
+          Official Microsoft Clarity Resources:
+        </h4>
+
+        <ul className="pl-[1.5rem] p-[0_0_23px_1em] list-disc">
+          <li className="leading-[26px] pl-[0.5rem] mb-[0.75rem] text-[14px] font-[500] text-[#333d47]">
+            <a href="https://clarity.microsoft.com/" className="text-[#0066B8] underline">
+              Microsoft Clarity Website
+            </a>
+          </li>
+          <li className="leading-[26px] pl-[0.5rem] mb-[0.75rem] text-[14px] font-[500] text-[#333d47]">
+            <a
+              href="https://learn.microsoft.com/en-us/clarity/"
+              className="text-[#0066B8] underline"
+            >
+              Microsoft Clarity Documentation
+            </a>
+          </li>
+        </ul>
+
+        <h4 className="text-[#1c2a38] text-[18px] pb-[10px] font-[700] leading-[23px] mt-[1rem]">
+          External Microsoft Clarity Resources:
+        </h4>
+        <ul className="p-[0_0_23px_1.5rem] list-disc">
+          <li className="underline pl-[0.5rem] mb-[0.75rem] text-[14px] leading-[26px]">
+            YouTube Clarity Tutorials:
+            <a
+              href="https://www.youtube.com/results?search_query=Microsoft+Clarity+Tutorial"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-[#0066B8] underline"
+            >
+              (Search for “Microsoft Clarity Tutorial” on YouTube.)
+            </a>
+          </li>
+          <li className="underline pl-[0.5rem] text-[#0066B8] mb-[0.75rem] text-[14px] leading-[26px]">
+            <a
+              href="https://neilpatel.com/blog/microsoft-clarity/"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-[#0066B8] underline"
+            >
+              Blog Posts on Website Analytics
+            </a>
+          </li>
+          <li className="underline pl-[0.5rem] text-[#0066B8] mb-[0.75rem] text-[14px] leading-[26px]">
+            <a
+              href="https://www.searchenginejournal.com/microsoft-clarity-case-studies/409608/"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-[#0066B8] underline"
+            >
+              Case Studies on Clarity
+            </a>
+          </li>
+          <li className="underline pl-[0.5rem] text-[#0066B8] mb-[0.75rem] text-[14px] leading-[26px]">
+            <a
+              href="https://techcommunity.microsoft.com/t5/microsoft-clarity-blog/bg-p/ClarityBlog"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-[#0066B8] underline"
+            >
+              Tech Community Discussions
+            </a>
+          </li>
+        </ul>
+      </section>
+    </>
+  )
+}
+
+export default Section7MicrosoftClarity
